插件优化

+++ yanye/occ/nckd-yanye-occ-plugin/src/main/java/nckd/yanye/occ/plugin/report/SaleInvOutDifReportListDataPlugin.java
@@ -34,14 +34,14 @@ public class SaleInvOutDifReportListDataPlugin extends AbstractReportListDataPlu
                 // 查询条件销售组织,标识如不一致,请修改
                 case "nckd_bizorg_q":
                     if (!(filterItem.getValue() == null)) {
-                        QFilter qFilter = new QFilter("bizorg", QCP.equals, (Long) ((DynamicObject) filterItem.getValue()).getPkValue());
+                        QFilter qFilter = new QFilter("bizorg", QCP.equals, ((DynamicObject) filterItem.getValue()).getPkValue());
                         qFilters.add(qFilter);
                     }
                     break;
                 // 查询条件销售部门,标识如不一致,请修改
                 case "nckd_bizdept_q":
                     if (!(filterItem.getValue() == null)) {
-                        QFilter qFilter = new QFilter("bizoperator", QCP.equals, (Long) ((DynamicObject) filterItem.getValue()).getPkValue());
+                        QFilter qFilter = new QFilter("bizoperator", QCP.equals, ((DynamicObject) filterItem.getValue()).getPkValue());
                         qFilters.add(qFilter);
                     }
                     break;
@@ -64,7 +64,7 @@ public class SaleInvOutDifReportListDataPlugin extends AbstractReportListDataPlu
                 // 查询条件物料,标识如不一致,请修改
                 case "nckd_material_q":
                     if (!(filterItem.getValue() == null)) {
-                        QFilter qFilter = new QFilter("billentry.material", QCP.equals, (Long) ((DynamicObject) filterItem.getValue()).getPkValue());
+                        QFilter qFilter = new QFilter("billentry.material", QCP.equals, ((DynamicObject) filterItem.getValue()).getPkValue());
                         qFilters.add(qFilter);
                     }
                     break;
@@ -98,7 +98,7 @@ public class SaleInvOutDifReportListDataPlugin extends AbstractReportListDataPlu
         List<Long> mainbillentryid = new ArrayList<>();
         while (copy.hasNext()) {
             Row next = copy.next();
-            if (next.getLong("mainbillentryid") != null && next.getLong("mainbillentryid")!= 0 ) {
+            if (next.getLong("mainbillentryid") != null && next.getLong("mainbillentryid")!= 0L ) {
                 mainbillentryid.add(next.getLong("mainbillentryid"));
             }
         }
